Docstrings

Forecast.Forecast — Module

Collection of methods for Time Series analysis

Methods implemented:

acf:        Auto-correlation or auto-covariance of univariate series. 
ccf:        Cross-correlation or cros-covariance of two univariate series.
d:          Lagged differences of a given order for Vector, Array and TimeSeries.
hma:        Henderson moving average filters.
loess:      Locally estimated scatterplot smoothing.
pacf:       Partial Auto-correlation function.
sma:        Simple moving average.
stl:        Seasonal and Trend decomposition using loess.

Forecast.CCF — Type

Package: Forecast

Store results from the functions acf, ccf and pacf

Arguments

`ccf::AbstractArray`    An array with results from ccf, acf and pacf
`N::Integer`            Length of ccf
`type::String`          Type of CCF
`lag::Integer`          Maximum number of lags
`alpha::Tuple`          CI thresholds
`ci::Tuple`             CI for alpha
`auto::Bool`            Auto-correlation
`call::String`          Method called to generate ccf

Forecast.STL — Type

Package: Forecast

Store results from the function stl

Arguments

`ta::TimeArray`    A time array with three time series from a fitted STL model
`call::String`     method called to generate ta

Forecast.acf — Method

Package: Forecast

acf(x::{AbstractVector,TimeArray},
    type = "cor",
    lag = Integer(ceil(10*log10(length(x)))),
    alpha = (0.95,0.99))

Compute the auto-correlation or auto-covariance for an univariate series.

The results are normalized to preserve homoscedasticity. The distribution used to normalize the data is an approximation of a Fisher Transformation via a Normal Distribution. There is a plot recipe for the returned object, if the type is cor the plot will also show confidence intervals for the given alpha values.

Arguments

x: Vector or uni-dimensional TimeArray of data.

type: Valid values are "cor" for correlation (default) and "cov" for convariance.

lag: Maximum number of lags.

alpha: A tuple with two thresholds (t1,t2) with t1 <= t2 to plot confidence intervals. The default values are 0.95 and 0.99.

Returns

A CCF object.

Examples

julia> x = rand(100);
res = acf(x1; type="cor");
plot(res)

Forecast.ccf — Method

Package: Forecast

ccf(x1::{AbstractVector,TimeArray},
    x2::{AbstractVector,TimeArray};
    type = "cor",
    lag = Integer(ceil(10*log10(length(x1)))),
    alpha = (0.95,0.99))

Compute the cross-correlation or cros-covariance of two univariate series.

The results are normalized to preserve homoscedasticity. The distribution used to normalize the data is an approximation of a Fisher Transformation via a Normal Distribution. There is a plot recipe for the returned object, if the type is cor the plot will also show confidence intervals for the given alpha values.

If, for a given integer k, x2 repeats x1 values such that x1[t] = x2[t+k] for all i then high correlation value will be placed at the right from the center in the results. That is, this convention will be represented in the plots as x1_t = x2_{t+k} -> _____0__k__ meaning x2 behavior can be predicted by x1 in k units.

Arguments

x1: Vector or uni-dimensional TimeArray of data.

x2: Vector or uni-dimensional TimeArray of data.

type: Valid values are "cor" for correlation (default) and "cov" for convariance.

lag: Maximum number of lags.

alpha: A tuple with two thresholds (t1,t2) with t1 <= t2 to plot confidence intervals. The default values are 0.95 and 0.99.

Returns

A CCF object.

Examples

julia> x1 = rand(100);
x2 = circshift(x1,6);
res = ccf(x1, x2; type="cor");
plot(res)

Forecast.co2 — Function

Package: Forecast

co2(full = false)

Return dataset with atmospheric Carbon Dioxide Dry Air Mole Fractions from quasi-continuous measurements at Mauna Loa, Hawaii.

K.W. Thoning, A.M. Crotwell, and J.W. Mund (2020), Atmospheric Carbon Dioxide Dry Air Mole Fractions from continuous measurements at Mauna Loa, Hawaii, Barrow, Alaska, American Samoa and South Pole. 1973-2019, Version 2020-08 National Oceanic and Atmospheric Administration (NOAA), Global Monitoring Laboratory (GML), Boulder, Colorado, USA https://doi.org/10.15138/yaf1-bk21 FTP path: ftp://aftp.cmdl.noaa.gov/data/greenhouse_gases/co2/in-situ/surface/

Arguments

full: if true Returns full original dataset from 1973 to 2020 in a DataFrame, otherwise returns the subset used in "STL: A Seasonal-Trend Decomposition Procedure Based on Loess" from Cleveland et. al. Its default value is false.

Returns

Dataframe or TimeArray containing the descrived dataset.

Examples

julia> co2()
[ Info: Dataset used in Cleveland et al. paper
4609×1 TimeArray{Union{Missing, Float64},1,Date,Array{Union{Missing, Float64},1}}
1974-05-17 to 1986-12-31
│            │ co2     │
├────────────┼─────────┤
│ 1974-05-17 │ 333.38  │
│ 1974-05-18 │ 333.11  │
│ 1974-05-19 │ 333.46  │
   [...]

Forecast.d — Function

Package: Forecast

function d(x::{AbstractVector,AbstractArray,TimeArray},
           order::Int=1;
           lag::Int=1,
           center::Bool=true,
           pad::Bool=true)

Return Lagged differences of a given order for Vector, Array and TimeSeries.

Arguments

x: Vector or Array of data.

order: Order of the differences; number of recursive iterations on the same vector/array.

lag: Lag for the difference.

center: Center the result in the response using Missing values.

pad: Includes or removes missing pad.

Returns

Lagged differences Vector or Array of a given order.

Examples

julia> x = [1,2,3,4,5];
julia> d(x)
5-element Array{Any,1}:
 1
 1
 1
 1
  missing

julia> d(x,2)
5-element Array{Any,1}:
  missing
 0
 0
 0
  missing

julia> d(x;lag=2,pad=false)
3-element Array{Any,1}:
 2
 2
 2

julia> x = reshape(collect(1:20),10,2)
10×2 Array{Int64,2}:
  1  11
  2  12
  3  13
  4  14
  5  15
  6  16
  7  17
  8  18
  9  19
 10  20 

julia> d(x,2;lag=2,pad=false)
6×2 Array{Any,2}:
 0  0
 0  0
 0  0
 0  0
 0  0
 0  0

julia> d(co2())
[ Info: Dataset used in Cleveland et al. paper
4609×1 TimeArray{Any,1,Date,Array{Any,1}} 1974-05-17 to 1986-12-31
│            │ A       │
├────────────┼─────────┤
│ 1974-05-17 │ -0.27   │
│ 1974-05-18 │ 0.35    │
│ 1974-05-19 │ 0.18    │
   ⋮
│ 1986-12-30 │ missing │
│ 1986-12-31 │ missing │

Forecast.hma — Method

Package: Forecast

hma(s, n)

Applies the Henderson moving average filter to dataset s with n-term.

Information about involved processes and application can be found at the following:

("Time Series Analysis: The Process of Seasonal Adjustment")[https://www.abs.gov.au/websitedbs/d3310114.nsf/4a256353001af3ed4b2562bb00121564/5fc845406def2c3dca256ce100188f8e!OpenDocument] Australian Bureau of Statistics

Arguments

s: Observations' support.

n: Observation values.

Returns

An array of Henderson filter smoothed values provided in s.

Examples

julia> hma(rand(24), 13)
24-element Array{Float64,1}:
[...]

Forecast.loess — Method

Package: Forecast

loess(xv,yv;
      d=2,
      q=Int64(round(3/4*length(xv))),
      rho=repeat([1.0],inner=length(xv)),  
      predict = xv)

Smooth a vector of observations using locally weighted regressions.

Although loess can be used to smooth observations for any given number of independent variables, this implementation is univariate. The speed of loess can be greatly increased by using fast aproximations for the linear fitting calculations, however this implementation calculates only exact results.

The loess functionality and nomenclature follows the descriptions in:

"STL: A Seasonal, Trend Decomposition Procedure Based on Loess" Robert B. Cleveland, William S. Cleveland, Jean E. McRae, and Irma Terpenning. Journal of Official Statistics Vol. 6. No. 1, 1990, pp. 3-73 (c) Statistics Sweden.

Arguments

xv: Observations' support.

yv: Observation values.

d: Degree of the linear fit, it accepts values 1 or 2.

q: As q increases loess becomes smoother, when q tends to infinity loess tends to an ordinary least square poynomial fit of degree d. It defaults to the rounding of 3/4 of xv's length.

rho: Weights expressing the reliability of the observations (e.g. if yi had variances sigma^2*ki where ki where known, the rhoi could be 1/ki). It defaults to 1.0.

predict: Vector containing the real values to be predicted, by default predicts xv.

Returns

The loess values for the values contained in predict.

Examples

julia> loess(rand(5), rand(5); predict=rand(10))
10-element Array{Float64,1}:
[...]

Forecast.nextodd — Method

Package: Forecast

nextodd(x)

Return the smallest odd integer greater than or equal to x.

Forecast.pacf — Method

Package: Forecast

pacf(x::{AbstractVector,TimeArray};
     type = "step-real",
     lag = Integer(ceil(10*log10(length(x1)))),
     alpha = (0.95,0.99))

Compute the partial auto-correlation for an univariate series.

There are two versions; the "step" version estimates the auto-regressive parameters of an increasing model, the "real" version estimates the actual partial auto-correlation by eliminating the linear information provided by the lags. When using the default type "stepwise-real" both versions are calculated.

The distribution used to estimate the confidence intervals is an approximation of a Fisher Transformation via a Normal Distribution. There is a plot recipe for the returned object.

Arguments

x: Vector or uni-dimensional TimeArray of data.

type = Valid values are "stepwise", "real" and "stepwise-real".

lag: Maximum number of lags.

alpha: A tuple with two thresholds (t1,t2) with t1 <= t2 to plot confidence intervals. The default values are 0.95 and 0.99.

Returns

A CCF object

Examples

julia> x = rand(100);
res = pacf(x);
plot(res)

Forecast.sma — Method

Package: Forecast

sma(x, n; center = true)

Smooth a vector of data using a simple moving average.

Arguments

x: Vector of data.

n: Size of the moving average.

center: centers the moving averaged values in the response.

Returns

Vector of moving average smoothed values containing missing values to preserve the size of the original vector.

Examples

julia> sma(1:5,3;center=true)
5-element Array{Any,1}:
  missing
 2.0
 3.0
 4.0
  missing

Forecast.stl — Method

Package: Forecast

stl(Yv, np; robust=false, 
            nl=nextodd(np), 
            ns=10*length(Yv)+1,
            nt=nextodd(1.5*np/(1-1.5/ns)), 
            ni=robust ? 1 : 2,
            no=0,
            spm=false,
            qsmp=max(div(np,7),2))

Decompose a time series into trend, seasonal, and remainder components.

"STL has a simple design that consists of a sequence of applications of the loess smoother; the simplicity allows analysis of the properties of the procedure and allows fast computation, even for very long time series and large amounts of trend and seasonal smoothing. Other features of STL are specification of amounts of seasonal and trend smoothing that range, in a nearly continous way, from very small amount of smoothing to a very large amount; robust estimates of the trend and seasonal components that are not distorted by aberrant behavior in the data; specification of the period of the seasonal component to any intenger multiple of the time sampling interval greater than one; and the ability to decompose time series with missing values."*

All default values are chosen following the recommendations of the original paper when those were recommended. ns is recommended to be chosen of the basis of knowledge of the time series and on the basis of diagnostic methods; it must nonethelessbe always odd and at least 7. A default value is not advised on the original paper, instead the same default value used in the stl implementation in R in usere here.

for no the authors advise 5 ("safe value") or 10 ("near certainty of convergence") cycles or a convergence criterion when robustness is required, in this case when robust is true computations stop when convergence is achieved in trend and seasonality.

for qsmp the authors do not adivise a default but they use a value close to div(np,7).

Arguments

np: Seasonality.

robust: Robust stl.

nl: Smoothing parameter of the low-pass filter.

ns: Smoothing parameter for the seasonal component.

nt: Smoothing parameter for the trend decomposition.

ni: Number of inner loop cycles.

no: Number of outer loop cycles.

spm: Seasonal post-smoothing.

qsmp: Loess q window for Seasonal post-smoothing.

verbose: If true shows updates for the Seasonal and Trend convergence.

cth: Corvengence threshold for Seasonal and Trend.

Returns

An STL object with the seasonal, trend and remainder components if Yv is an Array and a TimeSeries object with the same components if Yv is a TimeSeries.

STL: A Seasonal, Trend Decomposition Procedure Based on Loess" Robert B. Cleveland, William S. Cleveland, Jean E. McRae, and Irma Terpenning. Journal of Official Statistics Vol. 6. No. 1, 1990, pp. 3-73 (c) Statistics Sweden.

Examples

julia> stl(co2(),365; robust=true, spm=true)
[ Info: Dataset used in Cleveland et al. paper
[ Info: Corvengence achieved (< 0.01); Stopping computation...
STL{TimeSeries.TimeArray{Union{Missing, Float64},2,
Dates.Date,Array{Union{Missing, Float64},2}}}
(4609×3 TimeSeries.TimeArray{Union{Missing, Float64},2,
Dates.Date,Array{Union{Missing, Float64},2}} 1974-05-17 to 1986-12-31, 
"stl(Yn, np=365; nl=365, ns=46091, nt=549, ni=1, no=0, spm=true, qsmp=52)")