Methods

Auto-correlation/covariance function

Cros-correlation/covariance function

Lagged differences of a given order

Locally Estimated Scatterplot Smoothing (LOESS)

Seasonal and Trend decomposition based on Loess (STL)

Simple Moving Average

acf

Auto-correlation/covariance function

Forecast.acf — Function

Package: Forecast

acf(x::{AbstractVector,TimeArray},
    type = "cor",
    lag = Integer(ceil(10*log10(length(x)))),
    alpha = (0.95,0.99))

Compute the auto-correlation or auto-covariance for an univariate series.

The results are normalized to preserve homoscedasticity. The distribution used to normalize the data is an approximation of a Fisher Transformation via a Normal Distribution. There is a plot recipe for the returned object, if the type is cor the plot will also show confidence intervals for the given alpha values.

Arguments

x: Vector or uni-dimensional TimeArray of data.

type: Valid values are "cor" for correlation (default) and "cov" for convariance.

lag: Maximum number of lags.

alpha: A tuple with two thresholds (t1,t2) with t1 <= t2 to plot confidence intervals. The default values are 0.95 and 0.99.

Returns

A CCF object.

Examples

julia> x = rand(100);
res = acf(x1; type="cor");
plot(res)

ccf

Cros-correlation/covariance function

Forecast.ccf — Function

Package: Forecast

ccf(x1::{AbstractVector,TimeArray},
    x2::{AbstractVector,TimeArray};
    type = "cor",
    lag = Integer(ceil(10*log10(length(x1)))),
    alpha = (0.95,0.99))

Compute the cross-correlation or cros-covariance of two univariate series.

The results are normalized to preserve homoscedasticity. The distribution used to normalize the data is an approximation of a Fisher Transformation via a Normal Distribution. There is a plot recipe for the returned object, if the type is cor the plot will also show confidence intervals for the given alpha values.

If, for a given integer k, x2 repeats x1 values such that x1[t] = x2[t+k] for all i then high correlation value will be placed at the right from the center in the results. That is, this convention will be represented in the plots as x1_t = x2_{t+k} -> _____0__k__ meaning x2 behavior can be predicted by x1 in k units.

Arguments

x1: Vector or uni-dimensional TimeArray of data.

x2: Vector or uni-dimensional TimeArray of data.

type: Valid values are "cor" for correlation (default) and "cov" for convariance.

lag: Maximum number of lags.

alpha: A tuple with two thresholds (t1,t2) with t1 <= t2 to plot confidence intervals. The default values are 0.95 and 0.99.

Returns

A CCF object.

Examples

julia> x1 = rand(100);
x2 = circshift(x1,6);
res = ccf(x1, x2; type="cor");
plot(res)

d

Lagged differences of a given order

Forecast.d — Function

Package: Forecast

function d(x::{AbstractVector,AbstractArray,TimeArray},
           order::Int=1,
           lag::Int=1;
           center::Bool=false)

Return Lagged differences of a given order for Vector, Array and TimeSeries.

Arguments

x: Vector or Array of data.

order: Order of the differences; number of recursive iterations on the same vector/array.

lag: Lag for the difference.

center: Center the result in the response using Missing values.

Returns

Lagged differences Vector or Array of a given order.

Examples

julia> x = [1,2,3,4,5];
julia> d(x)
4-element Array{Int64,1}:
 1
 1
 1
 1

julia> d(x,2)
3-element Array{Int64,1}:
 0
 0
 0

julia> d(x,1,2)
3-element Array{Int64,1}:
 2
 2
 2

julia> x = reshape(collect(1:20),10,2)
10×2 Array{Int64,2}:
  1  11
  2  12
  3  13
  4  14
  5  15
  6  16
  7  17
  8  18
  9  19
 10  20 

julia> d(x,2,2)
8×2 Array{Any,2}:
 2  2
 2  2
 2  2
 2  2
 2  2
 2  2
 2  2
 2  2

julia> d(co2())
[ Info: Dataset used in Cleveland et al. paper
4608×1 TimeArray{Any,1,Date,Array{Any,1}} 1974-05-17 to 1986-12-30
│            │ A       │
├────────────┼─────────┤
│ 1974-05-17 │ -0.27   │
│ 1974-05-18 │ 0.35    │
│ 1974-05-19 │ 0.18    │
   ⋮
│ 1986-12-28 │ 0.04    │
│ 1986-12-29 │ -0.12   │
│ 1986-12-30 │ missing │

loess

Locally Estimated Scatterplot Smoothing (LOESS)

Forecast.loess — Function

Package: Forecast

loess(xv,yv;
      d=2,
      q=Int64(round(3/4*length(xv))),
      rho=repeat([1.0],inner=length(xv)),  
      predict = xv)

Smooth a vector of observations using locally weighted regressions.

Although loess can be used to smooth observations for any given number of independent variables, this implementation is univariate. The speed of loess can be greatly increased by using fast aproximations for the linear fitting calculations, however this implementation calculates only exact results.

The loess functionality and nomenclature follows the descriptions in:

"STL: A Seasonal, Trend Decomposition Procedure Based on Loess" Robert B. Cleveland, William S. Cleveland, Jean E. McRae, and Irma Terpenning. Journal of Official Statistics Vol. 6. No. 1, 1990, pp. 3-73 (c) Statistics Sweden.

Arguments

xv: Observations' support.

yv: Observation values.

d: Degree of the linear fit, it accepts values 1 or 2.

q: As q increases loess becomes smoother, when q tends to infinity loess tends to an ordinary least square poynomial fit of degree d. It defaults to the rounding of 3/4 of xv's length.

rho: Weights expressing the reliability of the observations (e.g. if yi had variances sigma^2*ki where ki where known, the rhoi could be 1/ki). It defaults to 1.0.

predict: Vector containing the real values to be predicted, by default predicts xv.

Returns

The loess values for the values contained in predict.

Examples

julia> loess(rand(5), rand(5); predict=rand(10))
10-element Array{Float64,1}:
[...]

p

Inverse lagged differences of a given order

Forecast.p — Function

Package: Forecast

function p(dx::{AbstractVector,AbstractArray,TimeArray},
           orderlag::{AbstractVector,AbstractArray} = [[0]])

Return reverse lagged differences of a given order for Vector, Array and TimeSeries.

Arguments

dx: Vector or Array of data.

orderlag: Initial values of the inverted series. The format per column for the order and lag values is [[lag1, lag2, ..., lagm], order2, order3, ..., ordern]. the default values is an initial constant of zero for an inversion of order one and lag one.

Returns

Lagged differences Vector or Array of a given order.

Examples

julia> x = repeat(1:12,3);
julia> dx = d(x,12,12);
julia> orderlag = vcat([collect(1:12)],repeat([0],11));
julia> p(dx,orderlag) ≈ x
true
julia> p(d(x),[[1]]) ≈ x
true

julia> tx = hcat(co2(), co2(), co2());
julia> vtx = values(tx);
julia> tx = TimeArray(timestamp(tx),replace(vtx, missing => 0.0));
julia> values(p(d(tx), [ [[333.38]] [[333.38]] [[333.38]] ] )) ≈ values(tx)
true
julia> values(p(d(tx,2), [ [[333.38]] [[333.38]] [[333.38]] ; -0.27 -0.27 -0.27] )) ≈ values(tx)
true

stl

Seasonal and Trend decomposition based on Loess

Forecast.stl — Function

Package: Forecast

stl(Yv, np; robust=false, 
            nl=nextodd(np), 
            ns=10*length(Yv)+1,
            nt=nextodd(1.5*np/(1-1.5/ns)), 
            ni=robust ? 1 : 2,
            no=0,
            spm=false,
            qsmp=max(div(np,7),2))

Decompose a time series into trend, seasonal, and remainder components.

"STL has a simple design that consists of a sequence of applications of the loess smoother; the simplicity allows analysis of the properties of the procedure and allows fast computation, even for very long time series and large amounts of trend and seasonal smoothing. Other features of STL are specification of amounts of seasonal and trend smoothing that range, in a nearly continous way, from very small amount of smoothing to a very large amount; robust estimates of the trend and seasonal components that are not distorted by aberrant behavior in the data; specification of the period of the seasonal component to any intenger multiple of the time sampling interval greater than one; and the ability to decompose time series with missing values."*

All default values are chosen following the recommendations of the original paper when those were recommended. ns is recommended to be chosen of the basis of knowledge of the time series and on the basis of diagnostic methods; it must nonethelessbe always odd and at least 7. A default value is not advised on the original paper, instead the same default value used in the stl implementation in R in usere here.

for no the authors advise 5 ("safe value") or 10 ("near certainty of convergence") cycles or a convergence criterion when robustness is required, in this case when robust is true computations stop when convergence is achieved in trend and seasonality.

for qsmp the authors do not adivise a default but they use a value close to div(np,7).

Arguments

np: Seasonality.

robust: Robust stl.

nl: Smoothing parameter of the low-pass filter.

ns: Smoothing parameter for the seasonal component.

nt: Smoothing parameter for the trend decomposition.

ni: Number of inner loop cycles.

no: Number of outer loop cycles.

spm: Seasonal post-smoothing.

qsmp: Loess q window for Seasonal post-smoothing.

verbose: If true shows updates for the Seasonal and Trend convergence.

cth: Corvengence threshold for Seasonal and Trend.

Returns

An STL object with the seasonal, trend and remainder components if Yv is an Array and a TimeSeries object with the same components if Yv is a TimeSeries.

STL: A Seasonal, Trend Decomposition Procedure Based on Loess" Robert B. Cleveland, William S. Cleveland, Jean E. McRae, and Irma Terpenning. Journal of Official Statistics Vol. 6. No. 1, 1990, pp. 3-73 (c) Statistics Sweden.

Examples

julia> stl(co2(),365; robust=true, spm=true)
[ Info: Dataset used in Cleveland et al. paper
[ Info: Corvengence achieved (< 0.01); Stopping computation...
STL{TimeSeries.TimeArray{Union{Missing, Float64},2,
Dates.Date,Array{Union{Missing, Float64},2}}}
(4609×3 TimeSeries.TimeArray{Union{Missing, Float64},2,
Dates.Date,Array{Union{Missing, Float64},2}} 1974-05-17 to 1986-12-31, 
"stl(Yn, np=365; nl=365, ns=46091, nt=549, ni=1, no=0, spm=true, qsmp=52)")